+++ src/app/workflow/launch-third-party/state/descriptors.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { ga4ghPath, ga4ghWorkflowIdPrefix } from '../../../shared/constants';
 import { Dockstore } from '../../../shared/dockstore.model';
 import { SourceFile } from '../../../shared/swagger';
-import { DescriptorsStore } from './descriptors-store.';
+import { DescriptorsStore } from './descriptors-store';
 
 @Injectable()
 export class DescriptorsService {
